BUG: Fix possible loss of data on type conversion
Fix possible loss of data on type conversion, typically from a (possibly
64-bit) `size_t` to a 32-bit integer type, as suggested by various
occurrences of VS2017 warning C4244 and C4267 (when these warnings are
not disabled):

> video\core\include\itkRingBuffer.hxx(193): warning C4244: 'initializing': conversion from 'itk::RingBuffer<itk::DataObject>::OffsetValueType' to 'unsigned int'

> Filtering\BiasCorrection\src\itkCompositeValleyFunction.cxx(26): warning C4267: 'initializing': conversion from 'size_t' to 'unsigned int'

> Numerics\Polynomials\src\itkMultivariateLegendrePolynomial.cxx(151): warning C4267: 'argument': conversion from 'size_t' to 'int'

> Core\Common\src\itkArrayOutputSpecialization.cxx(29): warning C4267: 'initializing': conversion from 'size_t' to 'unsigned int'

> core\common\include\itkNumericTraitsArrayPixel.h(126): warning C4267: 'return': conversion from 'size_t' to 'unsigned int'

> numerics\statistics\include\itkMeasurementVectorTraits.h(308): warning C4267: 'return': conversion from 'size_t' to 'itk::Statistics::MeasurementVectorTraits::MeasurementVectorLength'

> IO\MeshBase\include\itkMeshIOBase.h(386): warning C4244: 'return': conversion from 'const itk::MeshIOBase::SizeValueType' to 'unsigned int'

+++ Modules/Numerics/Polynomials/src/itkMultivariateLegendrePolynomial.cxx
@@ -125,8 +125,8 @@ void MultivariateLegendrePolynomial
 {
   if ( coefficients.size() != m_NumberOfCoefficients )
     {
-    throw CoefficientVectorSizeMismatch(static_cast<int>( coefficients.size() ),
-                                        static_cast<int>( m_NumberOfCoefficients) );
+    throw CoefficientVectorSizeMismatch(coefficients.size(),
+                                        m_NumberOfCoefficients);
     }
 
   // copy coefficients to array of double
